Create factory functions for weather objects
These factory functions will be used to create objects for storing
weather data. The reason for this rather than directly using the fetched
data is due to the ability for users to swap temperature units. If the
temperature isn't stored it's annoying to access and make the conversion
for each existing temperature. The alternative is to redo the fetch call
in the new unit but this incurs a network cost.